BE - weather file refactoring
