--- src/backend/evals/requirements.txt
@@ -1,4 +1,0 @@
-requests>=2.27.1
-Werkzeug>=2.0.3
-Flask>=2.0.3
-SQLAlchemy>=1.4.32
